refactor: extract SettlementAllocation into Allocation in its own module
Move the per-installment allocation dataclass out of settlement.py
into allocation.py and rename it from SettlementAllocation to
Allocation. The module path provides sufficient context.
