+++ money_warp/loan/__init__.py
@@ -1,16 +1,17 @@
 """Loan module for personal loan modeling with flexible payment schedules."""
 
+from .allocation import Allocation
 from .engines.interest_calculator import InterestCalculator, MoraStrategy
 from .installment import Installment
 from .loan import Loan
-from .settlement import AnticipationResult, Settlement, SettlementAllocation
+from .settlement import AnticipationResult, Settlement
 
 __all__ = [
+    "Allocation",
     "AnticipationResult",
     "Installment",
     "InterestCalculator",
     "Loan",
     "MoraStrategy",
     "Settlement",
-    "SettlementAllocation",
 ]
